% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_partitions}
\alias{create_partitions}
\title{Create partitions for a vector}
\usage{
create_partitions(vector_length, equal_parts = 100)
}
\arguments{
\item{vector_length}{The total length of the partition vector}

\item{equal_parts}{The size of each partition}
}
\value{
A vector of \code{vector_length} divided into \code{equal_parts} with possibly
one additional vector of size less than \code{equal_parts}
}
\description{
Given a vector length and the size of each partition, create a vector which
represents those partitions. There may be one more partition of size
less than the equally divided parts. For example, if the vector length is
14 and the desired partitions are size 3, then there will be 4 partitions
of length 3 and one of length 2: 1 1 1 2 2 2 3 3 3 4 4 4 5 5.
}
\examples{
create_partitions(14, 3)
create_partitions(10, 4)

}
