% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rank_response_ratio_summarize}
\alias{rank_response_ratio_summarize}
\title{Calculate rank response ratio summary}
\usage{
rank_response_ratio_summarize(
  df,
  effect_expr_thres = 0,
  p_expr_thres = 0.05,
  normalize = FALSE,
  bin_size = 5,
  separator = ";"
)
}
\arguments{
\item{df}{A dataframe with binding signals, effect expression, and p-values}

\item{effect_expr_thres}{A threshold for effect expression (default: 0)}

\item{p_expr_thres}{A threshold for p-values (default: 0.05)}

\item{normalize}{A boolean indicating whether to normalize the results (default: FALSE)}

\item{bin_size}{The number of records per group (partition)}

\item{separator}{A string separator used in the binding expression source string}
}
\value{
A list containing two dataframes: rr with response ratios calculated
for each group, and random with the random expectation values
}
\description{
Given a dataframe with binding signals, effect expression, and p-values,
calculate the rank response ratio for each group and summarize the results.
}
\examples{
# Given a small example dataframe similar to your input
df <- data.frame(
  experiment = rep(c("exp1", "exp2"), each = 6),
  source_expr = rep(c("src1", "src2"), times = 6),
  binding_signal = runif(6,0,.07),
  effect_expr = c(0.1, 0.2, -0.3, -0.4, 0.5, 0.6, 0.1,
                  0.2, -0.3, -0.4, 0.5, 0.6),
  p_expr = c(0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.01,
             0.02, 0.03, 0.04, 0.05, 0.06)
)
rank_response_ratio_summarize(df, bin_size = 3, separator = ";")

}
